pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "allianceauth-oidc-provider"
authors = [{name = "AaronKable", email = "aaronkable@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "allianceauth>=3.3.0",
    "django-oauth-toolkit>=2.3.0,<3.0.0",
]
dynamic = ["version", "description"]

[project.urls]
Homepage = "https://github.com/Solar-Helix-Independent-Transport/allianceauth-oidc-provider"
Source = "https://github.com/Solar-Helix-Independent-Transport/allianceauth-oidc-provider"
Tracker = "https://github.com/Solar-Helix-Independent-Transport/allianceauth-oidc-provider/issues"

[tool.flit.module]
name = "allianceauth_oidc"

[tool.isort]
profile = "django"
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DJANGO",
    "ESI",
    "FIRSTPARTY",
    "LOCALFOLDER"
]
known_esi = ["esi"]
known_django = ["django"]
skip_gitignore = true
